requirements.txt
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
docutils==0.12
elasticsearch==2.3.0
lockfile==0.12.2
lxml==3.6.0
nose==1.3.7
python-daemon==2.1.1
python-dateutil==2.5.3
PyYAML==3.11
requests==2.10.0
Requires==0.0.3
-e git+https://github.com/screepers/python-screeps.git@5fb556af10c6632d88c672fb88348964e4e11e97#egg=screepsapi
six==1.10.0
urllib3==1.15.1
websocket-client==0.37.0
